Refactored code for issues #64 and #75
