+++ src/main/java/com/simpleftp/filesystem/FileSystemUtils.java
@@ -1,0 +1,53 @@
+/*
+ *  Copyright (C) 2020  Edward Lynch-Milner
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.simpleftp.filesystem;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Calendar;
+
+/**
+ * This class provides utility methods to the FileSystsemPackage
+ */
+public final class FileSystemUtils {
+    /**
+     * Prevent instantiation
+     */
+    private FileSystemUtils() {}
+
+    /**
+     * The format for the date time string throughout the UI for files
+     */
+    public static final String FILE_DATETIME_FORMAT = "MMM dd HH:mm";
+
+    /**
+     * Parses the calendar object to the FILE_DATETIME_FORMAT
+     * @param calendar the calendar object to parse
+     * @return the formatted date time
+     */
+    public static String parseCalendarToFormattedDate(Calendar calendar) {
+        int hour = calendar.get(Calendar.HOUR_OF_DAY);
+        int minute = calendar.get(Calendar.MINUTE);
+        int month = calendar.get(Calendar.MONTH) + 1;
+        int day = calendar.get(Calendar.DAY_OF_MONTH);
+        int year = calendar.get(Calendar.YEAR);
+
+        LocalDateTime localDateTime = LocalDateTime.of(year, month, day, hour, minute);
+        return localDateTime.format(DateTimeFormatter.ofPattern(FILE_DATETIME_FORMAT));
+    }
+}
