+++ src/main/java/com/simpleftp/filesystem/LocalFile.java
@@ -17,9 +17,16 @@
 
 package com.simpleftp.filesystem;
 
+import com.simpleftp.filesystem.exceptions.FileSystemException;
 import com.simpleftp.filesystem.interfaces.CommonFile;
 
 import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.PosixFilePermission;
+import java.text.SimpleDateFormat;
+import java.util.Set;
 
 /**
  * A file representing a local file. While the interfaces specifies that some methods return a FileSystemException, that
@@ -88,5 +95,117 @@ public class LocalFile extends File implements CommonFile {
     @Override
     public long getSize() {
         return length();
+    }
+
+    /**
+     * Checks if POSIX permissions are supported
+     * @return the permissions, null if posix permissions are not supported
+     */
+    private String resolvePosixPermissions() {
+        Path path = toPath();
+
+        if (path.getFileSystem().supportedFileAttributeViews().contains("posix")) {
+            try {
+                Set<PosixFilePermission> permissionsSet = Files.getPosixFilePermissions(path);
+                String[] permissionsStringArr= new String[9]; // first 3 indices, owner rwx, next 3, group, last 3 others
+                for (int i = 0; i < 9; i++)
+                    permissionsStringArr[i] = "-";
+
+                for (PosixFilePermission permission : permissionsSet) {
+                    switch (permission) {
+                        case OWNER_READ: permissionsStringArr[0] = "r"; break;
+                        case OWNER_WRITE: permissionsStringArr[1] = "w"; break;
+                        case OWNER_EXECUTE: permissionsStringArr[2] = "x"; break;
+                        case GROUP_READ: permissionsStringArr[3] = "r"; break;
+                        case GROUP_WRITE: permissionsStringArr[4] = "w"; break;
+                        case GROUP_EXECUTE: permissionsStringArr[5] = "x"; break;
+                        case OTHERS_READ: permissionsStringArr[6] = "r"; break;
+                        case OTHERS_WRITE: permissionsStringArr[7] = "w"; break;
+                        case OTHERS_EXECUTE: permissionsStringArr[8] = "x"; break;
+                    }
+                }
+
+                StringBuilder permissions = new StringBuilder();
+
+                for (String s : permissionsStringArr)
+                    permissions.append(s);
+
+                return permissions.toString();
+            } catch (IOException e) {
+                return null;
+            }
+        } else {
+            return null;
+        }
+    }
+
+    /**
+     * Resolves permissions on systems that aren't posix standard. This can only return permissions for current user
+     * @return the permissions string
+     */
+    private String resolveNonPosixPermissions() {
+        String permissions = "";
+
+        if (canRead()) {
+            permissions += "r";
+        } else {
+            permissions += "-";
+        }
+
+        if (canWrite()) {
+            permissions += "-w";
+        } else {
+            permissions += "--";
+        }
+
+        if (canExecute()) {
+            permissions += "-x";
+        } else {
+            permissions += "--";
+        }
+
+        return permissions;
+    }
+
+    /**
+     * Calculates the permissions for a local file
+     * @return the permissions
+     */
+    private String calculateLocalPermissions() {
+        String permissions = resolvePosixPermissions();
+        if (permissions == null)
+            permissions = resolveNonPosixPermissions();
+
+        if (Files.isSymbolicLink(toPath())) {
+            permissions = "l" + permissions;
+        } else if (isDirectory()) {
+            permissions = "d" + permissions;
+        } else {
+            permissions = "-" + permissions;
+        }
+
+        return permissions;
+    }
+
+    /**
+     * Gets the permissions as a string in the unix form of ls command. For Non-posix systems, this just displays the permissions for the user running the program
+     *
+     * @return the permissions as a string
+     */
+    @Override
+    public String getPermissions() {
+        return calculateLocalPermissions();
+    }
+
+    /**
+     * Gets the modification time as a formatted String in the form of Month Day Hour:Minute, e.g Jan 01 12:50
+     *
+     * @return the formatted modification time String, can be null if can't be determined.
+     */
+    @Override
+    public String getModificationTime() {
+        long lastModified = lastModified();
+
+        return new SimpleDateFormat(FileSystemUtils.FILE_DATETIME_FORMAT).format(lastModified);
     }
 }
