+++ src/main/java/com/simpleftp/ui/UI.java
@@ -39,10 +39,7 @@ import org.apache.tika.Tika;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -223,7 +220,7 @@ public final class UI {
             // more exception types can be handled here
             if (ex instanceof FTPRemotePathNotFoundException || ex instanceof LocalPathNotFoundException) {
                 String path;
-                String headerPrefix = "";
+                String headerPrefix;
                 if (ex instanceof FTPRemotePathNotFoundException) {
                     path = ((FTPRemotePathNotFoundException)ex).getRemotePath();
                     headerPrefix = "Remote ";
@@ -422,22 +419,6 @@ public final class UI {
     public static boolean doFileSizeWarningDialog(String path) {
         FileSizeConfirmationDialog confirmationDialog = new FileSizeConfirmationDialog(path);
         return confirmationDialog.showAndGetConfirmation();
-    }
-
-    /**
-     * Takes a calendar object, presumably from a FTPFile and uses it to parse it to a time in the format for the UI to display
-     * @param calendar the calendar object to convert
-     * @return the formatted String
-     */
-    public static String parseCalendarToUITime(Calendar calendar) {
-        int hour = calendar.get(Calendar.HOUR_OF_DAY);
-        int minute = calendar.get(Calendar.MINUTE);
-        int month = calendar.get(Calendar.MONTH) + 1;
-        int day = calendar.get(Calendar.DAY_OF_MONTH);
-        int year = calendar.get(Calendar.YEAR);
-
-        LocalDateTime localDateTime = LocalDateTime.of(year, month, day, hour, minute);
-        return localDateTime.format(DateTimeFormatter.ofPattern(FILE_DATETIME_FORMAT));
     }
 
     /**
