+++ src/main/java/com/simpleftp/ui/files/FileLineEntry.java
@@ -17,21 +17,21 @@
 
 package com.simpleftp.ui.files;
 
+import com.simpleftp.filesystem.exceptions.FileSystemException;
 import com.simpleftp.filesystem.interfaces.CommonFile;
-import com.simpleftp.ftp.exceptions.FTPException;
 import com.simpleftp.ui.panels.FilePanel;
 
 /**
  * A line entry implementation for a File.
  */
-public class FileLineEntry extends LineEntry {
+final class FileLineEntry extends LineEntry {
 
     /**
      * Constructs a FileLineEntry
      * @param file the file that this entry represents. It is not this class' responsibility to ensure file is a normal file
      * @param owningPanel the panel owning this entry, i.e. the panel it is on
      */
-    public FileLineEntry(CommonFile file, FilePanel owningPanel) throws FTPException {
+    FileLineEntry(CommonFile file, FilePanel owningPanel) throws FileSystemException {
         super("file_icon.png", file, owningPanel);
     }
 }
