+++ src/main/java/com/simpleftp/ui/files/LineEntry.java
@@ -17,12 +17,9 @@
 
 package com.simpleftp.ui.files;
 
-import com.simpleftp.ftp.FTPSystem;
-import com.simpleftp.filesystem.LocalFile;
-import com.simpleftp.filesystem.RemoteFile;
+import com.simpleftp.filesystem.exceptions.FileSystemException;
 import com.simpleftp.filesystem.interfaces.CommonFile;
-import com.simpleftp.ftp.connection.FTPConnection;
-import com.simpleftp.ftp.exceptions.FTPException;
+import com.simpleftp.ftp.FTPSystem;
 import com.simpleftp.ui.UI;
 import com.simpleftp.ui.panels.FilePanel;
 import javafx.geometry.Pos;
@@ -36,16 +33,6 @@ import javafx.scene.text.Text;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.net.ftp.FTPFile;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.attribute.PosixFilePermission;
-import java.text.SimpleDateFormat;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Set;
 
 /**
  * This is an abstract class representing a line entry on the panel.
@@ -53,6 +40,8 @@ import java.util.Set;
  * Each file on the panel can be represented as an entry of the file listing.
  * Since all the info for the file is laid out as a line by line basis on the panel, this class will be called a LineEntry
  * It can represent a Directory or File Line entry
+ *
+ * To create a LineEntry for a File, pass in a CommonFile that is a normal file and for a Directory, a CommonFile that is a directory to LineEntry.newInstance
  */
 @Log4j2
 public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
@@ -82,12 +71,9 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
      * @param file the file this LineEntry represents
      * @param owningPanel the FilePanel this LineEntry is a part of
      */
-    protected LineEntry(String imageURL, CommonFile file, FilePanel owningPanel) throws FTPException {
+    protected LineEntry(String imageURL, CommonFile file, FilePanel owningPanel) throws FileSystemException {
         setSpacing(30);
         setStyle(UI.WHITE_BACKGROUND);
-        /**
-         * The pane containing the image and name
-         */
         HBox imageNamePanel = new HBox();
         imageNamePanel.setSpacing(30);
         image = new ImageView();
@@ -105,7 +91,7 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
     /**
      * Initialises the line entry
      */
-    protected void init() throws FTPException {
+    protected void init() throws FileSystemException {
         if (getChildren().size() > 1) {
             getChildren().clear();
         }
@@ -118,8 +104,7 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
             Tooltip.install(text, t);
         }
         text.setFont(Font.font("Monospaced"));
-        getChildren().add(image);
-        getChildren().add(text);
+        getChildren().addAll(image, text);
 
         setOnMouseEntered(e -> setStyle(UI.GREY_BACKGROUND_TRANSPARENT));
         setOnMouseExited(e -> setStyle(UI.WHITE_BACKGROUND));
@@ -140,240 +125,36 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
      * Retrieves the modification time and size string of the file. Note that if ftpConnection.getModificationTime fails, the file.getTimestamp is used even if it is not very accurate
      * @return modification time or Cannot be determined
      */
-    public String getModificationTimeAndSize() throws FTPException {
-        String modificationTime = "";
-        if (file instanceof LocalFile) {
-            LocalFile localFile = (LocalFile)file;
-            modificationTime = localFile.length() + " " + new SimpleDateFormat(UI.FILE_DATETIME_FORMAT).format(localFile.lastModified());
-        } else {
-            FTPConnection connection = owningPanel.getFileSystem().getFTPConnection();
-            connection = connection == null ? FTPSystem.getConnection():connection;
-            if (connection != null) {
-                RemoteFile remoteFile = (RemoteFile)file;
-                FTPFile ftpFile = remoteFile.getFtpFile();
-                String filePath = file.getFilePath();
-                String fileModTime = connection.getModificationTime(filePath);
-                if (fileModTime != null) {
-                    LocalDateTime dateTime = LocalDateTime.parse(fileModTime, DateTimeFormatter.ofPattern("HH:mm:ss dd/MM/yyyy"));
-                    modificationTime = dateTime.format(DateTimeFormatter.ofPattern(UI.FILE_DATETIME_FORMAT));
-                    modificationTime = (ftpFile.isValid() ? ftpFile.getSize():connection.getFileSize(filePath)) + " " + modificationTime;
-                } else {
-                    if (ftpFile.isValid()) {
-                        return ftpFile.getSize() + " " + UI.parseCalendarToUITime(ftpFile.getTimestamp());
-                    }
+    public String getModificationTimeAndSize() throws FileSystemException {
+        String modificationTime = file.getModificationTime();
+        if (modificationTime == null)
+            modificationTime = "Cannot be determined";
 
-                    log.warn("Could not determine modification time for file {}", filePath);
-                    return "Cannot be determined";
-                }
-            }
-        }
-
-        return modificationTime;
+        return file.getSize() + " " + modificationTime;
     }
 
     /**
      * Gets the file name with modification time size and permissions
      * @return String of file name, size, modification time, permissions
      */
-    private String getFileNameString() throws FTPException {
+    private String getFileNameString() throws FileSystemException {
         String fileName = file.getName();
-        String paddedName = ""; // the name that will be displayed on the UI.
+        StringBuilder paddedName; // the name that will be displayed on the UI.
 
         if (fileName.length() > FILE_NAME_LENGTH) {
-            paddedName = fileName.substring(0, FILE_NAME_LENGTH - 3) + "...";
+            paddedName = new StringBuilder(fileName.substring(0, FILE_NAME_LENGTH - 3) + "...");
         } else {
-            paddedName = fileName;
+            paddedName = new StringBuilder(fileName);
             for (int i = fileName.length(); i < FILE_NAME_LENGTH; i++) {
-                paddedName += " ";
+                paddedName.append(" ");
             }
         }
 
-        paddedName += "\t\t" + getModificationTimeAndSize() + " " + calculatePermissionsString();
+        String permissions = file.getPermissions();
+        permissions = permissions == null ? "N/A":permissions;
+        paddedName.append("\t\t").append(getModificationTimeAndSize()).append(" ").append(permissions);
 
-        return paddedName.trim();
-    }
-
-    /**
-     * Checks if POSIX permissions are supported
-     * @return the permissions, null if posix permissions are not supported
-     */
-    private String resolvePosixPermissions() {
-        Path path = ((LocalFile)file).toPath();
-
-        if (path.getFileSystem().supportedFileAttributeViews().contains("posix")) {
-            try {
-                Set<PosixFilePermission> permissionsSet = Files.getPosixFilePermissions(path);
-                String[] permissionsStringArr= new String[9]; // first 3 indices, owner rwx, next 3, group, last 3 others
-                for (int i = 0; i < 9; i++)
-                    permissionsStringArr[i] = "-";
-
-                for (PosixFilePermission permission : permissionsSet) {
-                    switch (permission) {
-                        case OWNER_READ: permissionsStringArr[0] = "r"; break;
-                        case OWNER_WRITE: permissionsStringArr[1] = "w"; break;
-                        case OWNER_EXECUTE: permissionsStringArr[2] = "x"; break;
-                        case GROUP_READ: permissionsStringArr[3] = "r"; break;
-                        case GROUP_WRITE: permissionsStringArr[4] = "w"; break;
-                        case GROUP_EXECUTE: permissionsStringArr[5] = "x"; break;
-                        case OTHERS_READ: permissionsStringArr[6] = "r"; break;
-                        case OTHERS_WRITE: permissionsStringArr[7] = "w"; break;
-                        case OTHERS_EXECUTE: permissionsStringArr[8] = "x"; break;
-                    }
-                }
-
-                String permissions = "";
-
-                for (String s : permissionsStringArr)
-                    permissions += s;
-
-                return permissions;
-            } catch (IOException e) {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-
-    /**
-     * Resolves permissions on systems that aren't posix standard. This can only return permissions for current user
-     * @return the permissions string
-     */
-    private String resolveNonPosixPermissions() {
-        LocalFile localFile = (LocalFile)file;
-        String permissions = "";
-
-        if (localFile.canRead()) {
-            permissions += "r";
-        } else {
-            permissions += "-";
-        }
-
-        if (localFile.canWrite()) {
-            permissions += "-w";
-        } else {
-            permissions += "--";
-        }
-
-        if (localFile.canExecute()) {
-            permissions += "-x";
-        } else {
-            permissions += "--";
-        }
-
-        return permissions;
-    }
-
-    /**
-     * Calculates the permissions for a local file
-     * @return the permissions
-     */
-    private String calculateLocalPermissions() {
-        LocalFile localFile = (LocalFile)file;
-
-        String permissions = resolvePosixPermissions();
-        if (permissions == null)
-            permissions = resolveNonPosixPermissions();
-
-        if (Files.isSymbolicLink(localFile.toPath())) {
-            permissions = "l" + permissions;
-        } else if (localFile.isDirectory()) {
-            permissions = "d" + permissions;
-        } else {
-            permissions = "-" + permissions;
-        }
-
-        return permissions;
-    }
-
-    /**
-     * Calculates the permissions for a remote file
-     * @return the permissions
-     */
-    private String calculateRemotePermissions() {
-        String permissions = "";
-        RemoteFile remoteFile = (RemoteFile)file;
-        FTPFile file = remoteFile.getFtpFile();
-
-        if (file.isSymbolicLink()) {
-            permissions += "l";
-        } else if (file.isDirectory()) {
-            permissions += "d";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.USER_ACCESS, FTPFile.READ_PERMISSION)) {
-            permissions += "r";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.USER_ACCESS, FTPFile.WRITE_PERMISSION)) {
-            permissions += "w";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.USER_ACCESS, FTPFile.EXECUTE_PERMISSION)) {
-            permissions += "x";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.GROUP_ACCESS, FTPFile.READ_PERMISSION)) {
-            permissions += "r";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.GROUP_ACCESS, FTPFile.WRITE_PERMISSION)) {
-            permissions += "w";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.GROUP_ACCESS, FTPFile.EXECUTE_PERMISSION)) {
-            permissions += "x";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.WORLD_ACCESS, FTPFile.READ_PERMISSION)) {
-            permissions += "r";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.WORLD_ACCESS, FTPFile.WRITE_PERMISSION)) {
-            permissions += "w";
-        } else {
-            permissions += "-";
-        }
-
-        if (file.hasPermission(FTPFile.WORLD_ACCESS, FTPFile.EXECUTE_PERMISSION)) {
-            permissions += "x";
-        } else {
-            permissions += "-";
-        }
-
-        return permissions;
-    }
-
-    /**
-     * Calculates the permissions for the file
-     * @return permissions string for the file
-     */
-    protected String calculatePermissionsString() {
-        String permissions = "";
-
-        if (file instanceof LocalFile) {
-            permissions = calculateLocalPermissions();
-        } else {
-            permissions = calculateRemotePermissions();
-        }
-
-        return permissions;
+        return paddedName.toString().trim();
     }
 
     /**
@@ -383,7 +164,6 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
     public String getFilePath() {
         return file.getFilePath();
     }
-
 
     @Override
     public int compareTo(LineEntry other) {
@@ -401,5 +181,27 @@ public abstract class LineEntry extends HBox implements Comparable<LineEntry> {
         }
         // we want to compare without extensions
         return n1.compareTo(n2);
+    }
+
+    /**
+     * Creates a new instance of LineEntry based on the file provided
+     * @param file the file for this line entry
+     * @param owningPanel the panel displaying this LineEntry
+     * @return the appropriate instance of LineEntry or null if file doesn't exist or an error occurs
+     */
+    public static LineEntry newInstance(CommonFile file, FilePanel owningPanel) {
+        try {
+            if (file.isADirectory())
+                return new DirectoryLineEntry(file, owningPanel);
+            else if (file.isNormalFile())
+                return new FileLineEntry(file, owningPanel);
+            else
+                return null;
+        } catch (FileSystemException ex) {
+            if (FTPSystem.isDebugEnabled())
+                ex.printStackTrace();
+
+            return null;
+        }
     }
 }
